firmware/logos/pyproject.toml
[project]
name = "logos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
]
